refactor: change structure of the project, refactor code js
