+++ coffee-house/javascript/variables/favorite.variable.js
@@ -1,18 +1,18 @@
 export const arrFavoriteCoffe = [
     {
-        img: "/coffee-house/images/favorite/coffee-slider-1.png",
+        img: "/coffee-house/assests/images/favorite/coffee-slider-1.png",
         name: "S’mores Frappuccino",
         description: "This new drink takes an espresso and mixes it with brown sugar and cinnamon before being topped with oat milk.",
         price: "$5.50",
     },
     {
-        img: "/coffee-house/images/favorite/coffee-slider-2.png",
+        img: "/coffee-house/assests/images/favorite/coffee-slider-2.png",
         name: "Caramel Macchiato",
         description: "Fragrant and unique classic espresso with rich caramel-peanut syrup, with the addition of delicate cream under whipped thick foam.",
         price: "$5.00",
     },
     {
-        img: "/coffee-house/images/favorite/coffee-slider-3.png",
+        img: "/coffee-house/assests/images/favorite/coffee-slider-3.png",
         name: "Ice coffee",
         description: "A popular summer drink that tones and invigorates. Prepared from coffee, milk and ice.",
         price: "$4.50",
