+++ coffee-house/javascript/variables/product.variable.js
