+++ coffee-house/scss/components/about.scss
@@ -2,14 +2,14 @@
 
 .about {
     height: 644px;
-    background-image: url(/coffee-house/video/videoCoffee.gif);
+    background-image: url(/coffee-house/assests/video/videoCoffee.gif);
     background-size: cover;
     border-radius: 40px;
     margin: 0px 40px;
     margin-top: 20px;
-    
+
     @include tablet {
-    background-position: 70% 70%;
+        background-position: 70% 70%;
 
     }
 
@@ -66,13 +66,13 @@
     align-items: center;
     gap: 8px;
     cursor: pointer;
-    position: relative; 
+    position: relative;
 
     &::after {
         content: "";
         width: 20px;
         height: 20px;
-        background-image: url("/coffee-house/images/about/coffee-cup.svg");
+        background-image: url("/coffee-house/assests/images/about/coffee-cup.svg");
         background-size: contain;
         background-repeat: no-repeat;
         position: absolute;
@@ -86,14 +86,14 @@
         display: flex;
     }
 
-    &:hover{
+    &:hover {
         padding-right: 24px;
     }
 }
 
 
 
-.about__block {
+.about__container {
     padding-top: 60px;
     padding-left: 60px;
     margin-top: 20px;
@@ -104,6 +104,6 @@
 }
 
 .about__button-img {
-        display: none;
-    
+    display: none;
+
 }
