+++ coffee-house/scss/components/favorite.scss
@@ -166,86 +166,6 @@
 
 }
 
-
-
-.favorite-coffee-leave-right {
-    display: flex;
-    animation: 0.71s leave ease;
-}
-
-
-
-.favorite-coffee-active-right {
-    display: flex;
-    animation: 0.7s active ease;
-}
-
-
-.favorite-coffee-leave-left {
-    display: flex;
-    animation: 0.71s leaveleft ease;
-}
-
-.favorite-coffee-active-left {
-    display: flex;
-    animation: 0.7s activeleft ease;
-}
-
-@keyframes leave {
-    from {
-        transform: translateX(0);
-        opacity: 1;
-        visibility: visible;
-    }
-
-    to {
-        transform: translateX(-100%);
-        opacity: 0;
-        visibility: hidden;
-    }
-}
-
-@keyframes active {
-    from {
-        transform: translateX(100%);
-        visibility: hidden;
-    }
-
-    to {
-        transform: translateX(0);
-        opacity: 1;
-        visibility: visible;
-    }
-}
-
-@keyframes leaveleft {
-    from {
-        transform: translateX(0);
-        opacity: 1;
-        visibility: visible;
-    }
-
-    to {
-        transform: translateX(100%);
-        opacity: 0;
-        visibility: hidden;
-    }
-}
-
-@keyframes activeleft {
-    from {
-        transform: translateX(-100%);
-        opacity: 0;
-        visibility: hidden;
-    }
-
-    to {
-        transform: translateX(0);
-        opacity: 1;
-        visibility: visible;
-    }
-}
-
 .favorite__coffee-container {
     display: flex;
     width: 480px;
