+++ coffee-house/scss/components/place.scss
@@ -25,7 +25,7 @@
 }
 
 .place__column:nth-child(1) .photo__bigger {
-    background-image: url(/coffee-house/images/place/place-1.png);
+    background-image: url(/coffee-house/assests/images/place/place-1.png);
     background-position: 40% 50%;
     border-radius: 20px;
     width: 660px;
@@ -39,7 +39,7 @@
 }
 
 .place__column:nth-child(2) .photo__bigger {
-    background-image: url(/coffee-house/images/place/place-2.png);
+    background-image: url(/coffee-house/assests/images/place/place-2.png);
     background-position: 40% 50%;
     border-radius: 20px;
     width: 660px;
@@ -53,7 +53,7 @@
 }
 
 .place__column:nth-child(1) .photo__smaller {
-    background-image: url(/coffee-house/images/place/place-4.png);
+    background-image: url(/coffee-house/assests/images/place/place-4.png);
     background-position: 40% 50%;
     border-radius: 20px;
     width: 660px;
@@ -65,7 +65,7 @@
 }
 
 .place__column:nth-child(2) .photo__smaller {
-    background-image: url(/coffee-house/images/place/place-3.png);
+    background-image: url(/coffee-house/assests/images/place/place-3.png);
     background-position: 40% 50%;
     border-radius: 20px;
     width: 660px;
