+++ index.html
@@ -5,9 +5,9 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="X-UA-Compatible" content="ie=edge">
-    <title>SH</title>
-    <script defer type="module" src="/coffee-house/javascript/burger.js"></script>
-    <script defer type="module" src="/coffee-house/javascript/slider.js"></script>
+    <title>Main</title>
+    <script defer type="module" src="/coffee-house/javascript/scripts/burger.js"></script>
+    <script defer type="module" src="/coffee-house/javascript/scripts/slider.js"></script>
     <link rel="stylesheet" href="/coffee-house/css/style.css">
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
@@ -18,7 +18,7 @@
 
 <body>
     <header class="header">
-        <img class="header__logo" src="/coffee-house/images/header/logo.svg" alt="company-logo">
+        <img class="header__logo" src="/coffee-house/assests/images/header/logo.svg" alt="Company">
 
         <ul class="header__pages">
             <li><a href="#favorite-coffee">Favorite coffee</a></li>
@@ -30,12 +30,13 @@
         <div class="header__menu">
 
             <a href="/coffee-house/menu/menu.html">Menu</a>
-            <img href="/coffee-house/menu/menu.html" src="/coffee-house/images/header/coffee-cup.svg" alt="menu-logo">
+            <img href="/coffee-house/menu/menu.html" src="/coffee-house/assests/images/header/coffee-cup.svg"
+                alt="menu-logo">
         </div>
 
         <button class="header__burger">
-            <img class="line" src="/coffee-house/images/header/burgerLine.svg" alt="burgerLine">
-            <img class="line" src="/coffee-house/images/header/burgerLine.svg" alt="burgerLine">
+            <img class="line" src="/coffee-house/assests/images/header/burgerLine.svg" alt="Line">
+            <img class="line" src="/coffee-house/assests/images/header/burgerLine.svg" alt="Line">
         </button>
 
         <div class="header__burger-menu">
@@ -60,15 +61,15 @@
                     Menu
                 </h2>
 
-                <img src="/coffee-house/images/header/coffee-cup.svg" alt="menu-logo">
+                <img src="/coffee-house/assests/images/header/coffee-cup.svg" alt="Logo">
 
             </a>
         </div>
 
-    </header>
+    </header> 
     <main>
         <section class="about">
-            <div class="about__block">
+            <div class="about__container">
 
                 <h2 id="about" class="about__title">
                     <span class="about__title-word">Enjoy</span> premium coffee at our charming cafe
@@ -91,7 +92,7 @@
 
         <section class="favorite">
 
-            <h2 class="favorite__title" id="favorite-coffee">
+            <h2 class="favorite__title">
                 Choose your <span class="favorite__title-word">favorite</span> coffee
             </h2>
 
@@ -220,7 +221,7 @@
                 </div>
 
             </div>
-            <img class="download__picture" src="/coffee-house/images/download/mobile-screens.png" alt="downloadPhone">
+            <img class="download__picture" src="/coffee-house/assests/images/download/mobile-screens.png" alt="Download">
 
         </section>
